requirements.txt
flask==2.3.3
python-dotenv==1.0.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.3.0
langchain==0.0.346
langchain-community==0.0.1
tiktoken==0.5.1